# docs/swagger.yaml
definitions:
  todo.TodoList:
    properties:
      activeAt:
        type: string
      completed:
        type: boolean
      id:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
  description: This is a sample server Todo API.
  termsOfService: http://swagger.io/terms/
  title: Todo API
  version: "1.0"
paths:
  /api/todo-list/tasks:
    get:
      description: Get tasks by their status (active or done)
      parameters:
      - default: active
        description: Status
        enum:
        - active
        - done
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/todo.TodoList'
            type: array
      summary: Find tasks by status
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task with title, active date and isCompleated fields
      parameters:
      - description: Task to create
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/todo.TodoList'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/todo.TodoList'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Task already exists
          schema:
            type: string
      summary: Create a new task
      tags:
      - tasks
  /api/todo-list/tasks/{id}:
    delete:
      description: Delete a task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Task not found
          schema:
            type: string
      summary: Delete a task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update an existing task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated task
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/todo.TodoList'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: Task not found
          schema:
            type: string
      summary: Update an existing task
      tags:
      - tasks
  /api/todo-list/tasks/{id}/done:
    put:
      description: Mark a task as done by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Task not found
          schema:
            type: string
      summary: Mark a task as done
      tags:
      - tasks
swagger: "2.0"